requirements.txt
flask==2.2.5
numpy==1.23.4
scipy==1.9.3
lxml==4.9.1
line-bot-sdk==2.3.0 
matplotlib
gspread==5.7.2
oauth2client==4.1.3
mpl_finance
arrow